requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
discord.py==2.3.2
oandapyV20==0.6.3
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.3
schedule==1.2.0
cryptography==41.0.7
